backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blis
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.1
cymem==2.0.11
distro==1.9.0
fastapi==0.115.8
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.10
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.12
numpy==2.2.3
openai==1.63.2
orjson==3.10.15
packaging==24.2
preshed==3.0.9
propcache==0.3.0
pydub
openai
psycopg2-binary==2.9.10
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy
spacy-legacy
spacy-loggers
SQLAlchemy==2.0.38
srsly==2.5.1
starlette==0.45.3
tenacity==9.0.0
thinc==8.3.4
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
